Move everything over to the double template implementations to verify everything works perfects, remove default template types where ambiguous and switch std=c++23 to c++20.

+++ src/world.hpp
@@ -16,11 +16,11 @@ class World
 public:
     World();
 
-    const std::vector<Primitive<float>>& getPrimitives() const;
+    const std::vector<Primitive<double>>& getPrimitives() const;
 
 private:
     unsigned int primitiveCount;
-    std::vector<Primitive<float>> primitives;
+    std::vector<Primitive<double>> primitives;
 };
